+++ CHANGELOG.md
@@ -11,8 +11,7 @@
   internal infrastructure and unpublished data. It is now the responsibility of
   other programs utlizing this library to provide masks, input TBs, etc.
 * Replace `BootstrapAlgError` in `get_linfit` with a logged warning. Use default
-  slope and offset values instead of failing when there are less than 125 valid
-  pixels.
+* Separate AMSR2 and AMSRE parameter coefficient specifications.
 
 # v0.2.0